requirements.txt
httplib2==0.11.3
lxml==4.6.2
plotly==4.14.3
pandas==1.4.1
rich==10.6.0
urllib3==1.26.8
ccxt==1.74.95
joblib==1.0.1
parse==1.19.0
pytest==6.2.2
pytest-mock==3.7.0
pytest-cov==2.12.1
reportlab==3.6.5
python-dotenv==0.19.2
matplotlib==3.3.4
reportlab==3.6.5
finta==1.3
ta==0.10.1
graphviz==0.20
gprof2dot==2021.2.21
python-dotenv==0.19.2
discord==1.7.3
